requirements.txt
torch
transformers==4.37.2
sentencepiece
protobuf==3.20.3
python-docx
streamlit
tqdm
https://github.com/VarunGumma/IndicTransToolkit/archive/refs/heads/main.zip